Fix external data properties
