+++ src/components/sections/SearchOptionsForExternalData.js
@@ -3,11 +3,11 @@ import { compose } from 'recompose';
 import * as Fields from '../../ui/components/Fields';
 import ValidatedField from '../Form/ValidatedField';
 import withDisabledAssetRequired from '../enhancers/withDisabledAssetRequired';
+import withMapFormToProps from '../enhancers/withMapFormToProps';
+import withExternalData from '../enhancers/withExternalData';
 import Section from './Section';
 import Row from './Row';
-import withDataSource from '../enhancers/withDataSource';
-import withFieldValues from './NameGeneratorListPrompts/withFieldValues';
-import withExternalDataPropertyOptions from './NameGeneratorListPrompts/withExternalDataPropertyOptions';
+import withExternalDataPropertyOptions from './SortOptionsForExternalData/withExternalDataPropertyOptions';
 
 const SearchOptions = ({
   maxAdditionalDisplayProperties,
@@ -65,9 +65,9 @@ const SearchOptions = ({
 export { SearchOptions };
 
 export default compose(
-  withDataSource,
+  withMapFormToProps(['dataSource']),
   withDisabledAssetRequired,
-  withFieldValues(['dataSource']),
+  withExternalData,
   withExternalDataPropertyOptions,
 )(SearchOptions);
